+++ packages/gasket-plugin-fastify/lib/index.js
@@ -61,7 +61,7 @@ module.exports = {
 
       const middlewares = [];
       await gasket.execApply('middleware', async (plugin, handler) => {
-        const middleware = handler(app);
+        const middleware = await handler(app);
         if (middleware) {
           if (middlewareConfig) {
             const pluginName = plugin && plugin.name ? plugin.name : '';